AI/blog_generator/requirements.txt
accelerate==1.10.1
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.7.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
Django==4.2.11
django-extensions==3.2.3
fastapi==0.116.1
filelock==3.19.1
fonttools==4.59.2
fsspec==2025.9.0
h11==0.16.0
httptools==0.6.4
huggingface-hub==0.34.4
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.6
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
polars==1.33.1
psutil==7.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
safetensors==0.6.2
scipy==1.16.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.4.4
starlette==0.47.3
sympy==1.14.0
timm==1.0.19
tokenizers==0.22.0
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.56.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2024.1
ultralytics==8.3.197
ultralytics-thop==2.0.17
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
haversine